server2/requirements.txt
# Server 2 — Intelligence Layer Dependencies
# All pinned to major version. Install: pip install -r requirements.txt

fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
supabase>=2.4,<3.0
pydantic>=2.6,<3.0
python-docx>=1.1,<2.0
PyPDF2>=3.0,<4.0
scikit-learn>=1.4,<2.0
numpy>=1.26,<3.0
httpx>=0.27,<1.0
crewai>=0.28,<1.0
python-multipart>=0.0.9,<1.0

# gzip is Python stdlib — no pip entry needed
# All secrets via Doppler — no python-dotenv
